Use Okapi 1.1.52

+++ OkapiSampleTests/ProjectConfig/DriverConfig.cs
@@ -5,7 +5,7 @@ namespace OkapiSampleTests.ProjectConfig
 {
     internal class DriverConfig : IDriverConfig
     {
-        public int TimeoutInSeconds => 10;
+        public double TimeoutInSeconds => 10;
         public DomUtilConfig SearchByAnchorConfig => null;
     }
 }
